Description
In this assignment, I have to generate a transcript for students. Information about the students, the courses that they have taken, and the grades that they have received for different activities of a course are stored in a text file named "input.txt". The program reads the text file and generates a transcript and write it to another text file.
EECS2030,3,1000,P10(90),P10(80),P30(60),E15(60),E15(44),E20(80),John is the format of the text file lines where [P|E] a number that shows the weight of the assessment and the number in the brackets is a number that shows the grade received for this assessment. In this format P stands for practical, which is associated to any types of assessment such as lab activity, exercises, homework, worksheets and so on. E stands for Exam.

Refer to UML for this project.